added backend for hints, moved forms into components, added markdown support for challenge descriptions and hints, fixed error when trying to delete challenge without files

+++ src/lib/database/challenges.ts
@@ -3,7 +3,7 @@
 import {prisma} from '@/src/lib/globals'
 import {Repository, readRepository} from "@/src/lib/database/repositories";
 import {Category} from "@prisma/client";
-import {deleteChallengeFile} from "@/src/lib/storage";
+import {deleteChallengeFile, readChallengeFiles} from "@/src/lib/storage";
 
 export type Challenge = {
     id: string
@@ -139,16 +139,12 @@ export async function updateChallenge(id: string, data: EditableChallenge) {
 }
 
 export async function deleteChallenge(id: string) {
-    const {repositoryId} = await prisma.challenge.findUniqueOrThrow({
-        where: {
-            id: id
-        },
-        select: {
-            repositoryId: true
-        }
-    })
+    const challenge = await readChallenge(id)
 
-    await deleteChallengeFile(`${repositoryId}/${id}`)
+    if (!challenge) return null
+
+    const files = await readChallengeFiles(challenge)
+    if (files.length > 0) await deleteChallengeFile(`${challenge.repository.id}/${id}`)
 
     const result = await prisma.challenge.delete({
         where: {
